various bug fixes
