+++ AutoDock/DisplayManager.swift
@@ -31,8 +31,13 @@ class DisplayManager: ObservableObject {
 			return
 		}
 
-		for screen in NSScreen.screens {
-			displays.append(screen)
+		displays = NSScreen.screens
+
+		// check if display parameters actually changed...
+		// this is to prevent the app from overriding the user
+		// if the Dock's autohide is toggled manually
+		if currentDisplays.map(\.localizedName) == displays.map(\.localizedName) {
+			return
 		}
 
 		if minWidthToShowDock == 0.0 {
@@ -44,53 +49,5 @@ class DisplayManager: ObservableObject {
 		}
 
 		currentDisplays = displays
-	}
-
-	private func toggleDockVisibility(hidden: Bool) {
-		// don't bother updating dock autohide setting if it matches the new value
-		let readTask = Process()
-		readTask.launchPath = "/usr/bin/defaults"
-		readTask.arguments = ["read", "com.apple.dock", "autohide"]
-
-		let pipe = Pipe()
-		readTask.standardOutput = pipe
-
-		do {
-			try readTask.run()
-			readTask.waitUntilExit()
-
-			let data = pipe.fileHandleForReading.readDataToEndOfFile()
-			if let output = String(data: data, encoding: .utf8) {
-				let trimmedOutput = output.trimmingCharacters(in: .whitespacesAndNewlines)
-				let dockIsCurrentlyHidden = trimmedOutput == "1"
-				print(dockIsCurrentlyHidden)
-				if hidden == dockIsCurrentlyHidden {
-					print("Dock not updated")
-					return
-				}
-			}
-		} catch {
-			print("Failed to read Dock settings: \(error)")
-		}
-
-		let writeTask = Process()
-		writeTask.launchPath = "/usr/bin/defaults"
-		writeTask.arguments = ["write", "com.apple.dock", "autohide", "-bool", hidden.description]
-
-		let killDockTask = Process()
-		killDockTask.launchPath = "/usr/bin/killall"
-		killDockTask.arguments = ["Dock"]
-
-		do {
-			try writeTask.run()
-			writeTask.waitUntilExit()
-
-			try killDockTask.run()
-			killDockTask.waitUntilExit()
-
-			print(hidden ? "Dock is now hidden." : "Dock is now visible.")
-		} catch {
-			print("Failed to update Dock settings: \(error)")
-		}
 	}
 }
